requirements.txt
geemap
streamlit
pandas
ee
streamlit-folium
